+++ refgenconf/exceptions.py
@@ -6,6 +6,8 @@ __all__ = ["DownloadJsonError", "GenomeConfigFormatError", "MissingAssetError",
            "MissingConfigDataError", "MissingGenomeError",
            "RefgenconfError", "UnboundEnvironmentVariablesError"]
 
+DOC_URL = "http://refgenie.databio.org/en/dev/genome_config/"
+
 
 class RefgenconfError(Exception):
     """ Base exception type for this package """
@@ -14,15 +16,18 @@ class RefgenconfError(Exception):
 
 class DownloadJsonError(RefgenconfError):
     """ Non-OK response from a JSON download attempt """
-    pass
+    def __init__(self, resp):
+        super(DownloadJsonError, self).__init__(
+            "No response provided" if resp is None else
+            "JSON: {}".format(resp.json()))
+        self.response = resp
 
 
 class GenomeConfigFormatError(RefgenconfError):
     """ Exception for invalid genome config file format. """
     def __init__(self, msg):
         spacing = " " if msg[-1] in ["?", "."] else "; "
-        suggest = "For config format documentation please see " \
-                  "http://refgenie.databio.org/en/dev/genome_config/"
+        suggest = "For config format documentation please see " + DOC_URL
         super(GenomeConfigFormatError, self).__init__(msg + spacing + suggest)